+++ NEWS.md
@@ -2,6 +2,7 @@
 
  * Fixes for DED export
  * Support legacy version DatabaseConnector v6.0
+ * Fix observation period overlap query
 
 # CdmOnboarding 4.1.0